move PasswordProvider and friends to a .secret package why: reflect the fact their use is no longer limited to the TLS domain
