+++ kroxylicious-api/src/main/java/io/kroxylicious/proxy/config/tls/KeyPair.java
@@ -10,6 +10,8 @@ import java.util.Objects;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 
+import io.kroxylicious.proxy.config.secret.PasswordProvider;
+
 import edu.umd.cs.findbugs.annotations.SuppressFBWarnings;
 
 /**
