+++ kroxylicious-api/src/test/java/io/kroxylicious/proxy/config/secret/InlinePasswordTest.java
@@ -4,7 +4,7 @@
  * Licensed under the Apache Software License version 2.0, available at http://www.apache.org/licenses/LICENSE-2.0
  */
 
-package io.kroxylicious.proxy.config.tls;
+package io.kroxylicious.proxy.config.secret;
 
 import org.junit.jupiter.api.Test;
