+++ kroxylicious-integration-test-support/src/main/java/io/kroxylicious/proxy/config/BuilderConfig.java
@@ -18,8 +18,8 @@ import io.sundr.builder.annotations.ExternalBuildables;
         "io.kroxylicious.proxy.config.admin.AdminHttpConfiguration",
         "io.kroxylicious.proxy.config.admin.EndpointsConfiguration",
         "io.kroxylicious.proxy.config.admin.PrometheusMetricsConfig",
-        "io.kroxylicious.proxy.config.tls.FilePassword",
-        "io.kroxylicious.proxy.config.tls.InlinePassword",
+        "io.kroxylicious.proxy.config.secret.FilePassword",
+        "io.kroxylicious.proxy.config.secret.InlinePassword",
         "io.kroxylicious.proxy.config.tls.Tls",
         "io.kroxylicious.proxy.config.tls.TrustStore",
         "io.kroxylicious.proxy.config.tls.InsecureTls",
